Ajax

+++ app/controllers/answers_controller.rb
@@ -2,23 +2,18 @@ class AnswersController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    @answer = question.answers.build(answer_params)
-
-    if @answer.save
-      redirect_to question_path(question), notice: 'Your answer successfully posted.'
-    else
-      render 'questions/show', locals: { question: question }
-    end
+    @answer = question.answers.new(answer_params)
+    @answer.author = current_user
+    @answer.save
   end
 
   def update
-    answer.update(answer_params)
-    redirect_to question_path(question)
+    answer.update(answer_params) if user_author?
+    @question = answer.question
   end
 
   def destroy
-    answer.destroy
-    redirect_to question_path(answer.question.id)
+    answer.destroy if user_author?
   end
 
   private
@@ -33,5 +28,9 @@ class AnswersController < ApplicationController
 
   def question
     @question ||= Question.find(params[:question_id])
+  end
+
+  def user_author?
+    current_user == answer.author
   end
 end
